% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotal.R
\name{pivotal}
\alias{pivotal}
\title{Pivotal quantities}
\usage{
pivotal(fit, n = 10000)
}
\arguments{
\item{fit}{an \code{\link{aov1r}} object}

\item{n}{number of simulations}
}
\value{
The simulations in a list.
}
\description{
Simulates from the generalized pivotal quantities.
}
