% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov1r.R
\name{predict.aov1r}
\alias{predict.aov1r}
\title{Prediction interval}
\usage{
\method{predict}{aov1r}(object, level = 0.95, ...)
}
\arguments{
\item{object}{an output of \code{\link{aov1r}}}

\item{level}{confidence level}

\item{...}{ignored}
}
\value{
A vector of length two, the bounds of the prediction interval.
}
\description{
Prediction interval for the one-way random effect ANOVA.
}
\examples{
dat <- simAOV1R(I=2, J=3, mu=10, sigmab=1, sigmaw=1)
fit <- aov1r(y ~ group, data=dat)
predict(fit)
}
