% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{simAOV1R}
\alias{simAOV1R}
\title{Simulation of one-way random effect ANOVA}
\usage{
simAOV1R(I, J, mu, sigmab, sigmaw)
}
\arguments{
\item{I}{integer, number of groups}

\item{J}{integer, number of replicates per group}

\item{mu}{numeric, overall mean}

\item{sigmab}{positive number, the between standard deviation}

\item{sigmaw}{positive number, the within standard deviation}
}
\value{
A dataframe.
}
\description{
Simulates a balanced one-way random effect ANOVA model.
}
\examples{
simAOV1R(I=2, J=3, mu=10, sigmab=1, sigmaw=1)
}
