% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotal.R
\name{pivotal}
\alias{pivotal}
\title{Generalized pivotal quantities}
\usage{
pivotal(fit, n = 10000)
}
\arguments{
\item{fit}{an \code{\link{aov1r}} object}

\item{n}{number of simulations}
}
\value{
The simulations in a list.
}
\description{
Simulates from the generalized pivotal quantities.
}
\examples{
dat <- simAOV1R(I=20, J=5, mu=10, sigmab=1, sigmaw=1)
fit <- aov1r(y ~ group, data=dat)
pivsims <- pivotal(fit)
pivsims$G_sigma2tot <- pivsims$G_sigma2b + pivsims$G_sigma2w
# Generalized confidence intervals:
lapply(pivsims, quantile, probs = c(0.025, 0.975))
# compare with the frequentist confidence intervals:
confint(fit, SDs = FALSE)
# Generalized prediction interval:
with(
  pivsims,
  quantile(rnorm(length(G_mu), G_mu, sqrt(G_sigma2tot)),
           probs = c(0.025, 0.975))
)
# compare with the frequentist prediction interval:
predict(fit)
}
